requirements.txt
flask
psycopg2-binary        # Postgres driver
SQLAlchemy             # ORM / DB layer
python-dotenv          # for config/env vars
requests               # for Kalshi API calls
nba_api                # for NBA game data
kalshi-python==2.1.4
cryptography           # Required by kalshi-python for authentication